JDBC connection check

+++ src/main/java/com/skhynix/untact/controller/UsrArticleController.java
@@ -15,14 +15,15 @@ public class UsrArticleController {
     @Autowired
     private ArticleService articleService;
 
-    @RequestMapping("usr/article/detail")
+    @RequestMapping("/usr/article/detail")
     @ResponseBody
     public Article showDetail(int id) {
         Article article = articleService.getArticle(id);
+
         return article;
     }
 
-    @RequestMapping("usr/article/list")
+    @RequestMapping("/usr/article/list")
     @ResponseBody
     public List<Article> showList(String searchKeywordType, String searchKeyword) {
         if (searchKeywordType != null) {
@@ -46,42 +47,47 @@ public class UsrArticleController {
 
     @RequestMapping("/usr/article/doAdd")
     @ResponseBody
-    public ResultData doAdd(String title, String body){
-
+    public ResultData doAdd(String title, String body) {
         if (title == null) {
-            return new ResultData("F-1","title을 입력해주세요.");
+            return new ResultData("F-1", "title을 입력해주세요.");
         }
+
         if (body == null) {
-            return new ResultData("F-1","body를 입력해주세요.");
+            return new ResultData("F-1", "body를 입력해주세요.");
         }
 
         return articleService.addArticle(title, body);
-
     }
 
     @RequestMapping("/usr/article/doDelete")
     @ResponseBody
-    public ResultData doDelete(int id) {
+    public ResultData doDelete(Integer id) {
+        if (id == null) {
+            return new ResultData("F-1", "id를 입력해주세요.");
+        }
+
         Article article = articleService.getArticle(id);
 
         if (article == null) {
             return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
         }
+
         return articleService.deleteArticle(id);
     }
 
-    @RequestMapping("/usr/article /doModify")
+    @RequestMapping("/usr/article/doModify")
     @ResponseBody
-    public ResultData doModify(Integer id, String title, String body){
+    public ResultData doModify(Integer id, String title, String body) {
         if (id == null) {
-            return new ResultData("F-1","id를 입력해주세요.");
+            return new ResultData("F-1", "id를 입력해주세요.");
         }
 
         if (title == null) {
-            return new ResultData("F-1","title을 입력해주세요.");
+            return new ResultData("F-1", "title을 입력해주세요.");
         }
+
         if (body == null) {
-            return new ResultData("F-1","body를 입력해주세요.");
+            return new ResultData("F-1", "body를 입력해주세요.");
         }
 
         Article article = articleService.getArticle(id);
@@ -91,6 +97,5 @@ public class UsrArticleController {
         }
 
         return articleService.modifyArticle(id, title, body);
-
     }
 }
